# specs/004-todo-app-features/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo App Enhancements API
  version: 1.0.0
  description: API for managing Todo tasks with enhanced features including due dates, priorities, search, filter, and bulk actions.

servers:
  - url: http://localhost:8000/api/v1
    description: Local Development Server

tags:
  - name: Tasks
    description: Operations related to Todo tasks

paths:
  /tasks:
    get:
      summary: Get a list of tasks
      description: Retrieve a list of all tasks, with optional search and filtering capabilities.
      operationId: get_tasks
      tags:
        - Tasks
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search term to filter tasks by title or description (case-insensitive).
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, completed]
          description: Filter tasks by their completion status.
        - in: query
          name: created_after
          schema:
            type: string
            format: date-time
          description: Filter tasks created after this timestamp.
        - in: query
          name: due_before
          schema:
            type: string
            format: date-time
          description: Filter tasks due before this timestamp.
        - in: query
          name: priority
          schema:
            type: string
            enum: [low, medium, high]
          description: Filter tasks by priority level.
        - in: query
          name: categories
          schema:
            type: array
            items:
              type: string
          description: Filter tasks by categories/tags.
      responses:
        "200":
          description: A list of tasks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        "500":
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a new task
      description: Create a new task with optional due date, priority, and categories.
      operationId: create_task
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        "201":
          description: Task created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/InternalServerError'

  /tasks/{task_id}:
    get:
      summary: Get a single task by ID
      description: Retrieve a specific task by its unique identifier.
      operationId: get_task_by_id
      tags:
        - Tasks
      parameters:
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
          description: Unique identifier of the task.
      responses:
        "200":
          description: The requested task.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update an existing task
      description: Update the details of an existing task.
      operationId: update_task
      tags:
        - Tasks
      parameters:
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
          description: Unique identifier of the task to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        "200":
          description: Task updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "400":
          $ref: '#/components/responses/BadRequest'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a task
      description: Delete a specific task by its unique identifier.
      operationId: delete_task
      tags:
        - Tasks
      parameters:
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
          description: Unique identifier of the task to delete.
      responses:
        "204":
          description: Task deleted successfully.
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

  /tasks/{task_id}/toggle-complete:
    patch:
      summary: Toggle task completion status
      description: Toggles the completion status of a specific task.
      operationId: toggle_task_completion
      tags:
        - Tasks
      parameters:
        - in: path
          name: task_id
          schema:
            type: string
            format: uuid
          required: true
          description: Unique identifier of the task to toggle.
      responses:
        "200":
          description: Task completion status toggled successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'

  /tasks/bulk-delete:
    post:
      summary: Bulk delete tasks
      description: Delete multiple tasks by their unique identifiers.
      operationId: bulk_delete_tasks
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_ids]
              properties:
                task_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
                  description: Array of unique identifiers for tasks to delete.
      responses:
        "204":
          description: Tasks deleted successfully.
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/InternalServerError'

  /tasks/bulk-toggle-complete:
    post:
      summary: Bulk toggle tasks completion status
      description: Toggles the completion status for multiple tasks by their unique identifiers.
      operationId: bulk_toggle_tasks_completion
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_ids, status]
              properties:
                task_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  minItems: 1
                  description: Array of unique identifiers for tasks to update.
                status:
                  type: string
                  enum: [pending, completed]
                  description: The target completion status for the selected tasks.
      responses:
        "200":
          description: Tasks completion status toggled successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/InternalServerError'

  /tasks/export:
    get:
      summary: Export tasks
      description: Export all tasks to JSON or CSV format.
      operationId: export_tasks
      tags:
        - Tasks
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [json, csv]
          required: true
          description: The desired export format.
      responses:
        "200":
          description: Tasks exported successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
            text/csv:
              schema:
                type: string
        "400":
          $ref: '#/components/responses/BadRequest'
        "500":
          $ref: '#/components/responses/InternalServerError'

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        due_date:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, completed]
          default: pending
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        categories:
          type: array
          items:
            type: string
          nullable: true
        is_recurring:
          type: boolean
          default: false
        recurrence_pattern:
          type: string
          enum: [daily, weekly, monthly]
          nullable: true
      required:
        - title

    TaskCreate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        due_date:
          type: string
          format: date-time
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        categories:
          type: array
          items:
            type: string
          nullable: true
        is_recurring:
          type: boolean
          default: false
        recurrence_pattern:
          type: string
          enum: [daily, weekly, monthly]
          nullable: true
      required:
        - title

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
        due_date:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [pending, completed]
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        categories:
          type: array
          items:
            type: string
          nullable: true
        is_recurring:
          type: boolean
          default: false
        recurrence_pattern:
          type: string
          enum: [daily, weekly, monthly]
          nullable: true

  responses:
    BadRequest:
      description: Invalid request payload or parameters.
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
    InternalServerError:
      description: An unexpected error occurred on the server.
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string